+++ src/ev-common.h
@@ -167,24 +167,15 @@ API_LOCAL void ev__ipc_init_frame_hdr(ev_ipc_frame_hdr_t* hdr,
 API_LOCAL void ev__loop_update_time(ev_loop_t* loop);
 
 /**
- * @brief Submit task to event loop without multi-thread support.
- * @note If need multi-thread support, use #ev__loop_submit_task_mt().
- * @param[in] loop      Event loop
- * @param[in] token     A pointer to the pending token
- * @param[in] cb        A callback when the pending task is active
- */
-API_LOCAL void ev__loop_submit_task(ev_loop_t* loop, ev_todo_t* token, ev_todo_cb cb);
-
-/**
  * @brief Submit task to event loop with multi-thread support.
  * @note MT-Safe
  * @note Use this function in threads that \p loop not running. If the thread
- *   has \p loop running, use #ev__loop_submit_task().
+ *   has \p loop running, use #ev_todo().
  * @param[in] loop  Event loop
  * @param[in] token Todo token
  * @param[in] cb    Callback
  */
-API_LOCAL void ev__loop_submit_task_mt(ev_loop_t* loop, ev_todo_t* token, ev_todo_cb cb);
+API_LOCAL void ev__loop_submit_task_mt(ev_loop_t* loop, ev_todo_token_t* token, ev_todo_cb cb);
 
 /**
  * @brief Get minimal length of specific \p addr type.